Colocando tratamento de excessões padrão e validação na latitude e longitude.

+++ src/main/java/br/com/playlist/weather/service/OpenWeatherMapService.java
@@ -6,5 +6,5 @@ public interface OpenWeatherMapService {
 
     double getTemperatureByCity(OpenWeatherConfig weatherConfig, String city);
 
-    double getTemperatureByLatLon(OpenWeatherConfig weatherConfig, int lat, int lon);
+    double getTemperatureByLatLon(OpenWeatherConfig weatherConfig, double lat, double lon);
 }
